Import/Export/Sanitization
The passing of info and functions between index.js and other scripts, in this case sanitize.cjs, is working. Data is passed to it, modified, and sent back. In that file, I can contain all the functions that make sure the user inputs are kosher; I'll start by making sure any selected defaults do not return with their node-terminal coloring hanging on.